Change so that setup.sh is run on wlan0 ifup to try and get around issues where it'd run before it had an internet connection

+++ setup.sh
@@ -1,12 +1,13 @@
 #!/bin/sh
 
-sed -ie '/^sh \/setup.sh/d' /etc/rc.local
-rm /setup.sh
+if [ $ACTION = "ifup" ] && [ $DEVICE = "wlan0" ]; then
+    rm /etc/hotplug.d/iface/40-setup
 
-if [[ -e /setup/install.sh ]]; then
-    cd /setup
-    chmod +x install.sh
-    sh install.sh
+    if [[ -e /setup/install.sh ]]; then
+        cd /setup
+        chmod +x install.sh
+        sh install.sh
+    fi
+
+    reboot
 fi
-
-reboot
